+++ Unosquare.FFME.Windows/Unosquare.FFME.Windows.csproj
@@ -66,8 +66,8 @@
     <Compile Include="Platform\PropertyMapper.cs" />
     <Compile Include="Platform\GuiContextType.cs" />
     <Compile Include="Rendering\ClosedCaptionsControl.cs" />
-    <Compile Include="Rendering\HostedImage.cs" />
-    <Compile Include="Rendering\HostedVisualElement.cs" />
+    <Compile Include="Rendering\ImageHost.cs" />
+    <Compile Include="Rendering\ElementHostBase.cs" />
     <Compile Include="Subtitles.cs" />
     <Compile Include="Platform\WindowsMediaConnector.cs" />
     <Compile Include="Platform\WindowsNativeMethods.cs" />
